Dev/song management

+++ backend/services/cloudinaryService.js
@@ -1,19 +1,23 @@
 const cloudinary = require('../configs/cloudinary');
 const fs = require('fs');
 
-const uploadToCloudinary = async (filePath, folder = 'uploads') => {
+const uploadToCloudinary = async (filePath, folder = 'uploads', options = {}) => {
   try {
     const result = await cloudinary.uploader.upload(filePath, {
-      folder: folder,
-      resource_type : "auto"
+      folder,
+      resource_type: options.resource_type || 'auto',
+      use_filename: true,
+      unique_filename: true,
     });
 
-    fs.unlinkSync(filePath); 
+    if (fs.existsSync(filePath)) {
+      fs.unlinkSync(filePath);
+    }
+
     return result;
   } catch (err) {
     throw new Error('Upload to Cloudinary failed: ' + err.message);
   }
 };
-
 
 module.exports = { uploadToCloudinary };
